react-redux-ify article list
